+++ R/get.R
@@ -1,12 +1,17 @@
 #' Get new files for a given state
 #'
-#' This function add new files to
-#' - your `R` directory.
-#' - your `tests` directory.
+#' @description
+#' This function adds new files to:
+#' - your `R` directory
+#' - your `testthat/tests` directory
+#'
+#' If `open` is `TRUE`, it will open the new files for editing.
 #'
 #' @inherit btt_reset_hard params return
 #' @param open `logical` specifies if new files are opened for interactive
 #'  editing
+#'
+#' @return `state`, invisibly; called for side effects.
 #'
 #' @examples \dontrun{
 #' btt_get("2.1.2")
@@ -21,52 +26,53 @@ btt_get <- function(state, open = rlang::is_interactive()) {
 
   # check out branch
   old_branch <- gert::git_branch(repo = path_repo())
-  gert::git_branch_checkout(state, repo = path_repo())
+
+  suppressMessages(
+    gert::git_branch_checkout(state, repo = path_repo())
+  )
+
+  path_template <- fs::path(path_repo(), "inst/template")
 
   # files to be copied
-  files_ussie_r <- repo_dir_ls("inst/template/R")
-  files_ussie_testthat <- repo_dir_ls("inst/template/testthat")
+  files_ussie <- NULL
+  if (fs::dir_exists(path_template)) {
+    files_ussie <- fs::dir_ls(path_template, recurse = TRUE, regexp = "\\.R$")
+  }
 
   # nothing to be copied
-  if (is.null(c(files_ussie_r, files_ussie_testthat))) {
-    cli::cli_alert_info("No files to be copied.")
+  if (length(files_ussie) == 0) {
+    cli::cli_alert_info("State {.val {state}} has no files to be copied.")
     return(invisible(state))
   }
 
   # TODO: list files to be copied
-  cli::cli_alert_warning("This will copy new files:")
+  cli::cli_alert_warning(
+    "State {.val {state}} will copy files to your package:"
+  )
+  files_ussie_rel <- fs::path_rel(files_ussie, start = path_template)
+  purrr::walk(
+    files_ussie_rel,
+    ~cli::cli_bullets(c("*" = "{.file {.x}}"))
+  )
   cli::cli_verbatim("")
 
   proceed <- yesno::yesno("Proceed?")
   if (!proceed) {
-    cli::cli_alert_danger("{.fn btt_get()} has been aborted.")
+    cli::cli_alert_danger("{.fn btt_get} has been aborted.")
     return(invisible(state))
   }
 
-  # copy R files
-  if (!is.null(files_ussie_r)) {
-    fs::dir_copy(
-      fs::path(path_repo(), "inst/template/R"),
-      root$find_file("R"),
-      overwrite = TRUE
-    )
-  }
-
-  # copy testthat files
-  if (!is.null(files_ussie_testthat)) {
-    fs::dir_copy(
-      fs::path(path_repo(), "inst/template/testthat"),
-      root$find_file("tests/testthat"),
-      overwrite = TRUE
-    )
-  }
+  # copy files
+  fs::dir_copy(path_template, root$find_file("."), overwrite = TRUE)
 
   cli::cli_alert_success("Copied template-files for state {.val {state}}.")
 
   # restore old branch
   gert::git_branch_checkout(old_branch, repo = path_repo())
 
-  # TODO: open new files for editing
+  # open new files for editing
+  purrr::walk(files_ussie_rel, usethis::edit_file, open = open)
+
 
   invisible(state)
 }
@@ -79,5 +85,5 @@ repo_dir_ls <- function(repo_path) {
     return(NULL)
   }
 
-  fs::dir_ls(path)
+  fs::dir_ls(path, recurse = TRUE, regexp = "\\.R$")
 }